Mrs Baylock - break up bloated utils
